requirements.txt
appnope==0.1.0
beautifulsoup4==4.6.0
certifi==2018.1.18
chardet==3.0.4
coolsms-python-sdk==2.0.3
decorator==4.2.1
Django==2.0.2
django-extensions==1.9.9
idna==2.6
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.1
lxml==4.1.1
parso==0.1.1
pexpect==4.3.1
pickleshare==0.7.4
Pillow==5.0.0
prompt-toolkit==1.0.15
psycopg2-binary==2.7.4
ptyprocess==0.5.2
Pygments==2.2.0
pygraphviz==1.3.1
python-magic==0.4.15
pytz==2018.3
requests==2.18.4
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
typing==3.6.4
urllib3==1.22
wcwidth==0.1.7
